add pidp11.sh and remove dependency on /run/pidp11
In the original code, when the POWER key (actually, the address select knob) is pressed,
if the HALT switch is up the simulator is supposed to reboot. The server does this by
invoking /opt/pidp11/bin/rebootsimh.sh or /opt/pidp11/bin/down.sh  The former writes
"exit" to /run/pidp11/tmpsimhcommand.txt ; the latter invokes "shutdown" to shut down the
Raspberry Pi (which assumes that the panel service is running as root). This has been
redefined so that pressing POWER with HALT down only terminates simh. The actual decision
is taken in simh itself; the panel server now only propagates the switch setting. If the
HALT switch is down, simh exits with a non-zero status.

This eliminates the dependence upon using a specific filename in /run/pidp11, so the
service shell script (/opt/pidp11/bin/pidp11.sh) now uses mktemp to generate a random
temporary file for the boot script and checks the exit status of simh to determine whether
to restart simh or to exit.